refactor : 헤더 SSE / 알림 컴포넌트에서 모든 API 통신 처리
