+++ src/shared/layout/Header.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
 import useWindowResize from "../../hooks/useWindowResize";
-import UserModal from "../../components/header/UserModal";
 import HeaderPc from "../../components/header/HeaderPc";
 import HeaderMobile from "../../components/header/HeaderMobile";
+import UserModal from "../../components/header/UserModal";
 import Chatroom from "../../components/chat/Chatroom";
 import { EventSourcePolyfill } from "event-source-polyfill";
 import { BACK_API } from "../../core/env";
@@ -10,13 +10,10 @@ import { getCookies, removeCookies } from "../../core/cookie";
 import { useNavigate } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
 import {
-  sendChatStatus,
   sendModalStatus,
+  sendChatStatus,
 } from "../../redux/modules/modal/modalSlice";
-import {
-  __patchAlarmState,
-  __sendAlarmModalStatus,
-} from "../../redux/modules/alarm/alarmSlice";
+import { __sendAlarmModalStatus } from "../../redux/modules/alarm/alarmSlice";
 import { __getChatList } from "../../redux/modules/chat/chatSlice";
 
 const Header = () => {
@@ -26,8 +23,8 @@ const Header = () => {
   const dispatch = useDispatch();
   const modalStatus = useSelector((state) => state.generalModal.toggleModal);
   const chatStatus = useSelector((state) => state.generalModal.toggleChat);
-  const EventSource = EventSourcePolyfill;
   const [alarmCount, setAlarmCount] = useState();
+  const EVENTSOURCE = EventSourcePolyfill;
 
   //알람 구독
   useEffect(() => {
@@ -35,18 +32,23 @@ const Header = () => {
       let eventSource;
       const fetchSSE = async () => {
         try {
-          eventSource = new EventSource(`${BACK_API}/alarm/subscribe`, {
+          eventSource = new EVENTSOURCE(`${BACK_API}/alarm/subscribe`, {
             headers: {
               Authorization: tokenValue,
             },
-            heartbeatTimeout: 60 * 1000 * 60,
+            heartbeatTimeout: 60 * 1000 * 60, //라이브러리 자체 기능인 45000초 관련 에러 안띄우는 방법1
           });
 
           /* EVENTSOURCE ONMESSAGE : 서버로부터 message를 수신했을 때 호출하는 이벤트 핸들러 */
           eventSource.onmessage = async (event) => {
             const res = await event.data;
-            //console.log(res);
+            //console.log("서버 응답 데이터 :", res);
             setAlarmCount(res);
+          };
+
+          /* EVENTSOURCE ONERROR : 에러가 발생하거나 EventSource 객체에서 error event가 감지되었을 때 호출하는 이벤트 핸들러 */
+          eventSource.onerror = async (event) => {
+            eventSource.close();
           };
 
           window.onbeforeunload = () => {
@@ -54,35 +56,24 @@ const Header = () => {
             //console.log("새로고침 이벤트 감지 후 eventSource 종료");
             eventSource.close();
           };
-
-          /* EVENTSOURCE ONERROR : 에러가 발생하거나 EventSource 객체에서 error event가 감지되었을 때 호출하는 이벤트 핸들러 */
-          eventSource.onerror = async (event) => {
-            eventSource.close();
-          };
         } catch (error) {}
       };
+
       fetchSSE();
       return () => eventSource.close();
     }
-  }, [tokenValue, EventSource]);
+  }, [tokenValue, EVENTSOURCE]);
 
-  //alarm 모달, 게시글 이동, 읽음 수정
+  //알람 & 채팅 모달
   const onClickAlarmModalHandler = () => {
     dispatch(__sendAlarmModalStatus(true));
   };
-  const onMoveSelectPageHandler = (postingId, alarmId) => {
-    dispatch(__sendAlarmModalStatus(false));
-    navigate(`/details/${postingId}`);
-    dispatch(__patchAlarmState(alarmId));
-  };
-
-  // 채팅 모달
   const onClickChatOpenHandler = () => {
     dispatch(__getChatList());
     dispatch(sendChatStatus(!chatStatus));
   };
 
-  //헤더 사람 아이콘 모달 내용 기능
+  //헤더 탭바 기능 유저 모달
   const onMoveMyProfileHandler = () => {
     navigate("/myprofile");
     dispatch(sendModalStatus(true));
@@ -106,7 +97,6 @@ const Header = () => {
           onInnerWidth={innerWidth}
           onAlarmCount={alarmCount}
           onClickAlarmModalHandler={onClickAlarmModalHandler}
-          onMoveSelectPageHandler={onMoveSelectPageHandler}
           onClickChatOpenHandler={onClickChatOpenHandler}
         />
       ) : (
@@ -114,7 +104,6 @@ const Header = () => {
           onInnerWidth={innerWidth}
           onAlarmCount={alarmCount}
           onClickAlarmModalHandler={onClickAlarmModalHandler}
-          onMoveSelectPageHandler={onMoveSelectPageHandler}
         />
       )}
 
@@ -132,6 +121,7 @@ const Header = () => {
           logoutClick={onMoveLogoutHandler}
         />
       )}
+
       {!chatStatus && <Chatroom />}
     </>
   );
